+++ user/src/lib.rs
@@ -3,7 +3,7 @@ use core::fmt::{self, Write};
 pub mod syscall;
 
 pub use syscall::{
-    close, exit, getdents, mount, open, read, stat, umount, write, Stat, O_RDONLY, O_WRONLY,
+    close, exit, getdents, mkdir, mount, open, read, stat, umount, write, Stat, O_RDONLY, O_WRONLY,
 };
 
 pub const STDIN: usize = 0;